requirements.txt
fastapi
uvicorn
pydantic
openai
python-dotenv
openai>=1.40.0
pyyaml>=6.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
jinja2>=3.1.4
numpy>=1.26.0
python-dotenv>=1.0.0
faster-whisper>=1.0.0
Django>=4.2,<5
streamlit>=1.30.0
pypdf>=4.0.0
